SEQNG-70 Replace `engine.Status` with `SequenceState` for `Sequence.status`
