+++ modules/edu.gemini.seqexec.engine/src/main/scala/edu/gemini/seqexec/engine/Step.scala
@@ -18,17 +18,8 @@ object Step {
     */
   def status(step: Step[Action \/ Result]): StepState = {
 
-    def errored(ar: Action \/ Result): Boolean =
-      ar match {
-        case (-\/(_)) => false
-        case (\/-(r)) => r match {
-          case Result.OK(_)    => false
-          case Result.Error(_) => true
-        }
-      }
-
     // At least an Action in this Step errored.
-    if (step.executions.isEmpty || step.any(errored)) StepState.Error("An action errored")
+    if (step.executions.isEmpty || step.any(Execution.errored)) StepState.Error("An action errored")
     // All actions in this Step are pending.
     else if (step.all(_.isLeft)) StepState.Pending
     // All actions in this Step were completed successfully.
