+++ modules/edu.gemini.seqexec.server/src/main/scala/edu/gemini/seqexec/server/SeqexecEngine.scala
@@ -86,12 +86,7 @@ class SeqexecEngine(settings: SeqexecEngine.Settings) {
       // TODO: Implement willStopIn
       SequenceView(seq.metadata, statusSequence(seq), engineSteps(seq), None)
 
-    private def statusSequence(seq: SequenceAR): SequenceState =
-      engine.Sequence.status(seq) match {
-          case engine.Status.Waiting   => SequenceState.Idle
-          case engine.Status.Completed => SequenceState.Completed
-          case engine.Status.Running   => SequenceState.Running
-        }
+    private def statusSequence(seq: SequenceAR): SequenceState = engine.Sequence.status(seq)
 
     private def engineSteps(seq: SequenceAR): List[Step] = {
